+++ sgt-back-end/package.json
@@ -2,10 +2,10 @@
   "name": "sgt-back-end",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start": "node index.js",
-    "dev:server": "nodemon index.js",
+    "start": "node app.js",
+    "dev:server": "nodemon app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
